Do not skip everything in _createNativeAnimation nativeView.layer is undefined
At least in some if not all cases (noted with ui-drawer), the layer will be undefined when `_createNativeAnimation` is called. All of the logic in this method was contained within the `if (nativeView?.layer) { }`.  However, the animation still needs the rest of the logic in this method. I found this by noting that the `_createNativeSpringAnimation` would work when the `_createNativeAnimation` would not work, and that the spring version of the animation does not care if the layer is undefined.